+++ serviceTest/src/test/scala/org/broadinstitute/dsde/workbench/dao/Google.scala
@@ -1,8 +1,11 @@
 package org.broadinstitute.dsde.workbench.dao
 
+import java.io.File
+
 import akka.actor.ActorSystem
 import org.broadinstitute.dsde.workbench.config.Config
-import org.broadinstitute.dsde.workbench.google.{HttpGoogleBigQueryDAO, HttpGoogleIamDAO, HttpGoogleStorageDAO}
+import org.broadinstitute.dsde.workbench.google.{GoogleCredentialModes, HttpGoogleBigQueryDAO, HttpGoogleIamDAO, HttpGoogleStorageDAO}
+import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
 
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutor}
 
@@ -12,7 +15,9 @@ object Google {
   lazy val system = ActorSystem()
   val ec: ExecutionContextExecutor = ExecutionContext.global
 
-  lazy val googleIamDAO = new HttpGoogleIamDAO(Config.GCS.qaEmail, Config.GCS.pathToQAPem, appName, metricBaseName)(system, ec)
-  lazy val googleBigQueryDAO = new HttpGoogleBigQueryDAO(appName, metricBaseName)(system, ec)
-  lazy val googleStorageDAO = new HttpGoogleStorageDAO(Config.GCS.qaEmail, Config.GCS.pathToQAPem, appName, metricBaseName)(system, ec)
+  val pemMode = GoogleCredentialModes.Pem(WorkbenchEmail(Config.GCS.qaEmail), new File(Config.GCS.pathToQAPem))
+
+  lazy val googleIamDAO = new HttpGoogleIamDAO(appName, pemMode, metricBaseName)(system, ec)
+  lazy val googleBigQueryDAO = new HttpGoogleBigQueryDAO(appName, pemMode, metricBaseName)(system, ec)
+  lazy val googleStorageDAO = new HttpGoogleStorageDAO(appName, pemMode, metricBaseName)(system, ec)
 }